Refactor project import code from editor.js. Removed dead code. Updated inline docs.
